Add files via upload
What is APRSdroid?
APRSdroid is an Android application for Amateur Radio operators. It allows reporting your position to the APRS (Automatic Packet Reporting System) network, displaying of nearby amateur radio stations and the exchange of APRS messages.
APRSdroid is Open Source Software written in Scala and licensed under the GPLv2.
You can contact the author via GitHub or on the XMPP chat-room aprsdroid@chat.yax.im (webchat). We used to have the IRC channel #aprsdroidon the FreeNode network, but that was killed in 2021.